requirements.txt
elasticsearch>=8.0.0,<9.0.0
faker
kafka-python-ng
netaddr
python-dotenv
